core-metastore crate
