--- crates/metastore/src/lib.rs
@@ -1,6 +1,0 @@
-pub mod error;
-pub mod metastore;
-pub mod models;
-
-pub use metastore::*;
-pub use models::*;
